Everything in float

+++ src/SelPlugins/DRMCairo/DRMCairo.h
@@ -23,7 +23,7 @@ struct SelDCSurface {
 	cairo_surface_t *surface;
 	cairo_t *cr;
 
-	int32_t w,h;	/* Keep surface size as it seems it's not possible to retrieve it*/
+	double w,h;	/* Keep surface size as it seems it's not possible to retrieve it*/
 };
 
 struct DCCard {
